# java/src/main/java/com/dsi/laboreos/dto/SeleccionarCampoRequest.java
package com.dsi.laboreos.dto;

public class SeleccionarCampoRequest {
    private String nombreCampo;

    public String getNombreCampo() {
        return nombreCampo;
    }

    public void setNombreCampo(String nombreCampo) {
        this.nombreCampo = nombreCampo;
    }
}